// src/app/models/Response/Venta/TB_Trasporte/Response-List-Trasporte.ts
export class ResponseListTrasporte {
    idTrasporte: number = 0;
    nombrePersona: string = "";
    tipoPersona: string = "";
    tipoDocumento: string = "";
    numeroDocumento: string = "";
    telefono: string = "";
    codigoUbigeo: string = "";
    direccion: string = "";
    numeroPlaca: string = "";
    licenciaConductor: string = "";
    tipoAutomovil: string = "";
}