// src/app/modules/models/Ventas/TB_Trasporte/Request-VW-Trasporte.ts
export class RequestVWTrasporte {
    idTrasporte: number = 0;
    nombrePersona: string = "";
    tipoPersona: string = "";
    tipoDocumento: string = "";
    numeroDocumento: string = "";
    telefono: string = "";
    codigoUbigeo: string = "";
    direccion: string = "";
    numeroPlaca: string = "";
    licenciaConductor: string = "";
    tipoAutomovil: string = "";
}